♻️ Refactor: 게임 초기값 세팅 리팩토링
